Added information for Warp to Warp with MWAN (#22883)
* Added information for Warp to Warp with MWAN

We have it documented that MWAN and Warp connectors are unsupported, but theres also specific considerations needed for using Warp to Warp when MWAN is in use. Added a note to expand on this.

* Update warp.mdx

Cleaned up note and removed the more technical reasons on how the change works and why its needed.

* Update src/content/docs/magic-wan/zero-trust/warp.mdx

---------

Co-authored-by: marciocloudflare <83226960+marciocloudflare@users.noreply.github.com>